Remove support for mirrors (API, db, etc.) (#147)
We are not relying on mirrors for quite some time now. I could find any client that uses the `mirror_list/` API endpoint in the mozilla, mozilla-services, and mozilla-releng orgs. Therefore this seems _safe_. If there is a client that uses this endpoint, we should change it anyway because go-bouncer does not use the `mirror_mirrors` table.